Add files via upload
This is an auto clicker I made that uses pixel detection to find the golden cookie for the game cookie clicker. It's an offline game, so no real harm is done. Just felt like doing something to cheat the system a bit. Ask me if you're having issues running the program. I'm sure I can help. Issue is that this is lacking the img files for the golden cookie, so it physically won't run unless you 1, understand my code, or 2, as me for some help.